+++ frb_codegen/src/library/codegen/parser/mod.rs
@@ -10,7 +10,6 @@ mod sanity_checker;
 pub(crate) mod source_graph;
 pub(crate) mod type_alias_resolver;
 pub(crate) mod type_parser;
-mod unused_checker;
 
 use crate::codegen::dumper::Dumper;
 use crate::codegen::ir::func::IrFunc;
@@ -25,11 +24,11 @@ use crate::codegen::parser::function_parser::FunctionParser;
 use crate::codegen::parser::internal_config::ParserInternalConfig;
 use crate::codegen::parser::misc::parse_has_executor;
 use crate::codegen::parser::reader::CachedRustReader;
-use crate::codegen::parser::sanity_checker::check_suppressed_input_path_no_content;
+use crate::codegen::parser::sanity_checker::misc_checker::check_suppressed_input_path_no_content;
 use crate::codegen::parser::source_graph::modules::Struct;
 use crate::codegen::parser::type_alias_resolver::resolve_type_aliases;
 use crate::codegen::parser::type_parser::TypeParser;
-use crate::codegen::parser::unused_checker::get_unused_types;
+use crate::codegen::parser::sanity_checker::unused_checker::get_unused_types;
 use crate::codegen::ConfigDumpContent;
 use crate::library::misc::consts::HANDLER_NAME;
 use anyhow::ensure;
